new files: vote and tests
